🥅 chore: Comment 컴포넌트 모킹 데이터 제거
